misc: functions for casting (#469)
* misc: add function to cast uint256 to shorter type

* misc: add tests

* misc: add casting to controller functions

* fix oracle and wPowerPerp extra address conversion

* misc: change mint type to uint256

* misc: update revert message

* misc: naming and small optimzations

+++ packages/hardhat/contracts/interfaces/IController.sol
@@ -17,13 +17,13 @@ interface IController {
 
     function mintPowerPerpAmount(
         uint256 _vaultId,
-        uint128 _powerPerpAmount,
+        uint256 _powerPerpAmount,
         uint256 _uniTokenId
     ) external payable returns (uint256 vaultId, uint256 wPowerPerpAmount);
 
     function mintWPowerPerpAmount(
         uint256 _vaultId,
-        uint128 _wPowerPerpAmount,
+        uint256 _wPowerPerpAmount,
         uint256 _uniTokenId
     ) external payable returns (uint256 vaultId);
